requirements.txt
streamlit>=1.36.0
chromadb>=0.5.5
sentence-transformers>=3.0.1
pandas>=2.2.2
numpy>=1.26.4
pydantic>=2.7.3
python-dotenv>=1.0.1
scikit-learn>=1.5.0
ragas>=0.1.13
datasets>=2.20.0
evaluate>=0.4.2
